src/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "fastapi>=0.115.12",
    "playwright>=1.50.0",
    "requests>=2.32.3",
    "uvicorn>=0.34.2",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "joblib>=1.3.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9", # For PostgreSQL
    "passlib[bcrypt]>=1.7.4", # For password hashing
    "python-jose[cryptography]>=3.3.0", # For JWT
    "python-multipart>=0.0.9", # For FastAPI form data (login)
]
